[Fix] #82 - 체험 & 예약 BigInteger -> Long으로 수정

+++ baro-support/src/main/java/com/barofarm/support/experience/presentation/dto/ReservationRequest.java
@@ -5,7 +5,6 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
@@ -45,7 +44,7 @@ public class ReservationRequest {
     @Schema(description = "총 금액", example = "60000", requiredMode = Schema.RequiredMode.REQUIRED)
     @NotNull(message = "총 금액은 필수입니다")
     @Min(value = 0, message = "총 금액은 0원 이상이어야 합니다")
-    private BigInteger totalPrice;
+    private Long totalPrice;
 
     /** Command DTO를 Service DTO로 변환 */
     public ReservationServiceRequest toServiceRequest() {
