+++ baro-support/src/test/java/com/barofarm/support/experience/application/ReservationServiceTest.java
@@ -2,27 +2,15 @@ package com.barofarm.support.experience.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.Mockito.*;
 
 import com.barofarm.support.common.client.FarmClient;
 import com.barofarm.support.common.exception.CustomException;
 import com.barofarm.support.experience.application.dto.ReservationServiceRequest;
 import com.barofarm.support.experience.application.dto.ReservationServiceResponse;
-import com.barofarm.support.experience.domain.Experience;
-import com.barofarm.support.experience.domain.ExperienceRepository;
-import com.barofarm.support.experience.domain.ExperienceStatus;
-import com.barofarm.support.experience.domain.Reservation;
-import com.barofarm.support.experience.domain.ReservationRepository;
-import com.barofarm.support.experience.domain.ReservationStatus;
+import com.barofarm.support.experience.domain.*;
 import com.barofarm.support.experience.exception.ReservationErrorCode;
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.Arrays;
 import java.util.Optional;
@@ -70,18 +58,18 @@ class ReservationServiceTest {
         reservationId = UUID.randomUUID();
         farmId = UUID.randomUUID();
 
-        validRequest = new ReservationServiceRequest(experienceId, buyerId, LocalDate.of(2025, 3, 15), "10:00-12:00", 2,
-                BigInteger.valueOf(30000), null);
+        validRequest = new ReservationServiceRequest(experienceId, buyerId, LocalDate.of(2025, 3, 15),
+            "10:00-12:00", 2, 30000L, null);
 
         validReservation = new Reservation(reservationId, experienceId, buyerId, LocalDate.of(2025, 3, 15),
-                "10:00-12:00", 2, BigInteger.valueOf(30000), ReservationStatus.REQUESTED);
+                "10:00-12:00", 2, 30000L, ReservationStatus.REQUESTED);
 
         // Experience는 protected 생성자이므로 Reflection을 사용하거나 Mock을 사용
         // 예약 날짜(2025-03-15)가 체험 가능 기간 내에 있도록 설정
         java.time.LocalDateTime availableStart = java.time.LocalDateTime.of(2025, 3, 1, 9, 0);
         java.time.LocalDateTime availableEnd = java.time.LocalDateTime.of(2025, 3, 31, 18, 0);
         experience = new Experience(experienceId, farmId, "Test Experience", "Description",
-                BigInteger.valueOf(15000), 20, 120, availableStart, availableEnd,
+                15000L, 20, 120, availableStart, availableEnd,
                 com.barofarm.support.experience.domain.ExperienceStatus.ON_SALE);
     }
 
@@ -149,7 +137,7 @@ class ReservationServiceTest {
         UUID sellerId = UUID.randomUUID();
         UUID reservationId2 = UUID.randomUUID();
         Reservation reservation2 = new Reservation(reservationId2, experienceId, buyerId, LocalDate.of(2025, 3, 16),
-                "14:00-16:00", 3, BigInteger.valueOf(45000), ReservationStatus.CONFIRMED);
+                "14:00-16:00", 3, 45000L, ReservationStatus.CONFIRMED);
 
         Pageable pageable = PageRequest.of(0, 10);
         Page<Reservation> reservationPage = new PageImpl<>(
@@ -177,7 +165,7 @@ class ReservationServiceTest {
         // given
         UUID otherExperienceId = UUID.randomUUID();
         Reservation reservation2 = new Reservation(UUID.randomUUID(), otherExperienceId, buyerId,
-                LocalDate.of(2025, 3, 16), "14:00-16:00", 1, BigInteger.valueOf(15000), ReservationStatus.REQUESTED);
+                LocalDate.of(2025, 3, 16), "14:00-16:00", 1, 15000L, ReservationStatus.REQUESTED);
 
         Pageable pageable = PageRequest.of(0, 10);
         Page<Reservation> reservationPage = new PageImpl<>(
@@ -271,7 +259,7 @@ class ReservationServiceTest {
     void createReservation_ExperienceNotAvailable() {
         // given
         Experience closedExperience = new Experience(experienceId, UUID.randomUUID(), "Test Experience", "Description",
-                BigInteger.valueOf(15000), 20, 120, java.time.LocalDateTime.now(), java.time.LocalDateTime.now().plusDays(30),
+                15000L, 20, 120, java.time.LocalDateTime.now(), java.time.LocalDateTime.now().plusDays(30),
                 ExperienceStatus.CLOSED);
         when(experienceRepository.findById(experienceId)).thenReturn(Optional.of(closedExperience));
 
@@ -291,7 +279,7 @@ class ReservationServiceTest {
         // given
         java.time.LocalDate invalidDate = LocalDate.of(2025, 6, 1); // 체험 가능 기간 밖
         ReservationServiceRequest invalidDateRequest = new ReservationServiceRequest(experienceId, buyerId, invalidDate,
-                "10:00-12:00", 2, BigInteger.valueOf(30000), null);
+                "10:00-12:00", 2, 30000L, null);
         when(experienceRepository.findById(experienceId)).thenReturn(Optional.of(experience));
 
         // when & then
@@ -347,7 +335,7 @@ class ReservationServiceTest {
         // given
         UUID sellerId = UUID.randomUUID();
         Reservation canceledReservation = new Reservation(reservationId, experienceId, buyerId, LocalDate.of(2025, 3, 15),
-                "10:00-12:00", 2, BigInteger.valueOf(30000), ReservationStatus.CANCELED);
+                "10:00-12:00", 2, 30000L, ReservationStatus.CANCELED);
         when(reservationRepository.findById(reservationId)).thenReturn(Optional.of(canceledReservation));
         // 상태 검증이 먼저 수행되므로 experienceRepository와 farmClient stubbing 불필요
 
@@ -366,7 +354,7 @@ class ReservationServiceTest {
         // given
         // CANCELED는 구매자만 가능하므로 buyerId 사용
         Reservation completedReservation = new Reservation(reservationId, experienceId, buyerId, LocalDate.of(2025, 3, 15),
-                "10:00-12:00", 2, BigInteger.valueOf(30000), ReservationStatus.COMPLETED);
+                "10:00-12:00", 2, 30000L, ReservationStatus.COMPLETED);
         when(reservationRepository.findById(reservationId)).thenReturn(Optional.of(completedReservation));
 
         // when & then
@@ -384,7 +372,7 @@ class ReservationServiceTest {
     void deleteReservation_CannotDeleteConfirmedOrCompleted() {
         // given
         Reservation confirmedReservation = new Reservation(reservationId, experienceId, buyerId, LocalDate.of(2025, 3, 15),
-                "10:00-12:00", 2, BigInteger.valueOf(30000), ReservationStatus.CONFIRMED);
+                "10:00-12:00", 2, 30000L, ReservationStatus.CONFIRMED);
         when(reservationRepository.findById(reservationId)).thenReturn(Optional.of(confirmedReservation));
 
         // when & then
@@ -416,7 +404,7 @@ class ReservationServiceTest {
     void deleteReservation_CanDeleteCanceled() {
         // given
         Reservation canceledReservation = new Reservation(reservationId, experienceId, buyerId, LocalDate.of(2025, 3, 15),
-                "10:00-12:00", 2, BigInteger.valueOf(30000), ReservationStatus.CANCELED);
+                "10:00-12:00", 2, 30000L, ReservationStatus.CANCELED);
         when(reservationRepository.findById(reservationId)).thenReturn(Optional.of(canceledReservation));
         doNothing().when(reservationRepository).deleteById(reservationId);
